+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
 * dev-master
-    * HOTFIX      #- Fixed executions null variable after deserialize
+    * HOTFIX      #16 Added sort-order to storage interface
+    * HOTFIX      #-  Fixed executions null variable after deserialize
 
 * 0.2.0 (2016-02-27)
     * ENHANCEMENT #7 Added task execution to hold information for one single try